paneles separados
se separaron los paneles del pedido
