+++ src/main/java/com/itec1/e_commerce/views/resources/JCarrierFieldListener.java
@@ -4,7 +4,7 @@ import com.itec1.e_commerce.controllers.OrderPanelController;
 import java.util.List;
 import javax.swing.event.DocumentEvent;
 import javax.swing.event.DocumentListener;
-import com.itec1.e_commerce.views.NewOrder_SecondPanel;
+import com.itec1.e_commerce.views.NewOrder_ChooseCarrierPanel;
 
 /**
  *
@@ -14,9 +14,9 @@ public class JCarrierFieldListener<T, U> implements DocumentListener {
 
     private List<T> entities;
     private final OrderPanelController controller;
-    private final NewOrder_SecondPanel panel;
+    private final NewOrder_ChooseCarrierPanel panel;
 
-    public JCarrierFieldListener(List<T> entities, OrderPanelController controller, NewOrder_SecondPanel panel) {
+    public JCarrierFieldListener(List<T> entities, OrderPanelController controller, NewOrder_ChooseCarrierPanel panel) {
         this.entities = entities;
         this.controller = controller;
         this.panel = panel;
@@ -24,16 +24,16 @@ public class JCarrierFieldListener<T, U> implements DocumentListener {
 
     @Override
     public void changedUpdate(DocumentEvent e) {
-        entities = (List<T>) controller.updateCarrierTable((panel.getCarrierFilter()));
+        entities = (List<T>) controller.updateCarrierTableByTransportType((panel.getCarrierFilter()));
     }
 
     @Override
     public void insertUpdate(DocumentEvent e) {
-        entities = (List<T>) controller.updateCarrierTable((panel.getCarrierFilter()));
+        entities = (List<T>) controller.updateCarrierTableByTransportType((panel.getCarrierFilter()));
     }
 
     @Override
     public void removeUpdate(DocumentEvent e) {
-        entities = (List<T>) controller.updateCarrierTable(panel.getCarrierFilter());
+        entities = (List<T>) controller.updateCarrierTableByTransportType(panel.getCarrierFilter());
     }
 }
