+++ src/main/java/com/itec1/e_commerce/views/resources/JWarehouseTableFieldListener.java
@@ -1,10 +1,10 @@
 package com.itec1.e_commerce.views.resources;
 
 import com.itec1.e_commerce.controllers.OrderPanelController;
+import com.itec1.e_commerce.views.NewOrder_ChooseWarehousePanel;
 import java.util.List;
 import javax.swing.event.DocumentEvent;
 import javax.swing.event.DocumentListener;
-import com.itec1.e_commerce.views.NewOrder_SecondPanel;
 
 /**
  *
@@ -14,9 +14,9 @@ public class JWarehouseTableFieldListener<T, U> implements DocumentListener {
 
     private List<T> entities;
     private final OrderPanelController controller;
-    private final NewOrder_SecondPanel panel;
+    private final NewOrder_ChooseWarehousePanel panel;
 
-    public JWarehouseTableFieldListener(List<T> entities, OrderPanelController controller, NewOrder_SecondPanel panel) {
+    public JWarehouseTableFieldListener(List<T> entities, OrderPanelController controller, NewOrder_ChooseWarehousePanel panel) {
         this.entities = entities;
         this.controller = controller;
         this.panel = panel;
